basic/tfrecord/requirements.txt
tensorflow==2.12.0
numpy>=1.19.2
pyyaml